Update test.html
Added features We added an info window to display additional information about each traffic incident when the user clicks on a marker. 	•	We included the Google Maps Traffic Layer to show real-time traffic conditions on the map. 	•	We added a search box that allows users to search for specific locations or addresses on the map using the Google Places API.